moved download into journalist_app/utils.py
